% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\docType{package}
\name{cpdb}
\alias{cpdb}
\title{cpdb : Package for residue contacts of a PDB solved structure
The `cpdb` package provides one important function, namely, `get_contacts()`.}
\section{`cpdb` functions}{

The `get_contacts` function computes the contacts of a giver reisude position.
The `plot_closest` plots the closest residues to the target given a distance in angstroms.
}

