% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts}
\alias{get_contacts}
\title{Compute distance matrix}
\usage{
get_contacts(pdb, ch = "A", res, a = 8)
}
\arguments{
\item{pdb}{PDB name ID string}

\item{ch}{PDB chain string (default "A")}

\item{res}{Residue position}

\item{a}{Distance in angstroms (default 8 angstroms)}
}
\value{
A list with the residue contacts.
}
\description{
Compute distance matrix
}
\examples{
x <- get_contacts("4q21", "A", 15, 6)
y <- get_contacts("1XI4", "B", 312)



}
