% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_closest.R
\name{plot_closest}
\alias{plot_closest}
\title{Plot residues at certain distance (angstroms) from specified residue}
\usage{
plot_closest(pdb, ch = "A", res, a = 8)
}
\arguments{
\item{pdb}{PDB name ID string}

\item{ch}{PDB chain string (default "A")}

\item{res}{Residue position}

\item{a}{Distance in angstroms (default 8 angstroms)}
}
\value{
A scatter plot where x = distance to specified residue.
}
\description{
Plot residues at certain distance (angstroms) from specified residue
}
\examples{
x <- plot_closest("4q21", "A", 15, 6)
y <- plot_closest("1XI4", "B", 312)



}
